update OrderDataContainer constructor

+++ Engine/src/OrderProduct.java
@@ -4,7 +4,7 @@ public class OrderProduct extends StoreProduct
 {
     private float amount;
 
-    public OrderProduct(StoreProduct storeProduct, int amount)
+    public OrderProduct(StoreProduct storeProduct, float amount)
     {
         super(storeProduct);
         this.amount = amount;
